Changes to be committed: 	modified:   .env.example 	deleted:    app/Http/Controllers/Recursos/TelefonoController.php 	new file:   app/Http/Controllers/Resources/ComputerController.php 	new file:   app/Http/Controllers/Resources/TelephoneController.php 	modified:   app/Http/Controllers/rrhh/UsersController.php 	new file:   app/Resources/Computer.php 	renamed:    app/Recursos/Telefono.php -> app/Resources/Telephone.php 	modified:   app/User.php 	renamed:    database/factories/RecursoTelefonoFactory.php -> database/factories/RecursosComputerFactory.php 	new file:   database/factories/RecursosTelephoneFactory.php 	new file:   database/migrations/2018_02_20_153950_create_computers_table.php 	renamed:    database/migrations/2018_02_16_152224_create_telefonos_table.php -> database/migrations/2018_02_20_154119_create_telephones_table.php 	modified:   database/seeds/DatabaseSeeder.php 	deleted:    database/seeds/RecursoTelefonosTableSeeder.php 	new file:   database/seeds/ResourcesComputersTableSeeder.php 	new file:   database/seeds/ResourcesTelephonesTableSeeder.php 	modified:   resources/views/layouts/partials/nav.blade.php 	deleted:    resources/views/recursos/telefono/directorio.blade.php 	deleted:    resources/views/recursos/telefono/index.blade.php 	new file:   resources/views/resources/computer/create.blade.php 	new file:   resources/views/resources/computer/edit.blade.php 	new file:   resources/views/resources/computer/index.blade.php 	renamed:    resources/views/recursos/telefono/create.blade.php -> resources/views/resources/telephone/create.blade.php 	new file:   resources/views/resources/telephone/directory.blade.php 	renamed:    resources/views/recursos/telefono/edit.blade.php -> resources/views/resources/telephone/edit.blade.php 	new file:   resources/views/resources/telephone/index.blade.php 	modified:   resources/views/rrhh/edit.blade.php 	modified:   routes/web.php

+++ app/Http/Controllers/rrhh/UsersController.php
@@ -97,8 +97,10 @@ class UsersController extends Controller
      */
     public function edit(User $user)
     {
+        $cargos = \App\rrhh\Cargo::orderBy('name','asc')->get();
         return view('rrhh.edit')
-            ->with('user',$user);
+            ->with('user',$user)
+            ->with('cargos',$cargos);
     }
 
     /**
@@ -110,8 +112,18 @@ class UsersController extends Controller
      */
     public function update(Request $request, User $user)
     {
-        // Detach all cargos from the user...
-        $user->cargos()->detach();
+        if ($request->has('cargos')) {
+            if ($request->filled('cargos')) {
+                foreach($request->input('cargos') as $key=>$cargo){
+                    $user->cargos()->attach($cargo);
+                }
+            }
+            else {
+                // Detach all cargos from the user...
+                $user->cargos()->detach();
+            }
+        }
+        
 
         $user->fill($request->all());
         $user->save();
